+++ webknossos/webknossos/cli/export_wkw_as_tiff.py
@@ -269,7 +269,7 @@ def main(
     logging.info("Starting tiff export for bounding box: %s", bbox)
     executor_args = Namespace(
         jobs=jobs,
-        distribution_strategy=distribution_strategy,
+        distribution_strategy=distribution_strategy.value,
         job_resources=job_resources,
     )
     used_tile_size = None