+++ webknossos/webknossos/dataset/dataset.py
@@ -37,6 +37,7 @@ from upath import UPath
 
 from webknossos.dataset.defaults import (
     DEFAULT_CHUNK_SHAPE,
+    DEFAULT_CHUNKS_PER_SHARD_FROM_IMAGES,
     DEFAULT_CHUNKS_PER_SHARD_ZARR,
 )
 
@@ -769,8 +770,6 @@ class Dataset:
         Creates a new layer called `layer_name` and adds it to the dataset.
         The dtype can either be specified per layer or per channel.
         If neither of them are specified, `uint8` per channel is used as default.
-        When creating a "Segmentation Layer" (`category="segmentation"`),
-        the parameter `largest_segment_id` also has to be specified.
 
         Creates the folder `layer_name` in the directory of `self.path`.
 
@@ -837,10 +836,6 @@ class Dataset:
             self._properties.data_layers += [layer_properties]
             self._layers[layer_name] = Layer(self, layer_properties)
         elif category == SEGMENTATION_CATEGORY:
-            assert (
-                "largest_segment_id" in kwargs
-            ), f"Failed to create segmentation layer {layer_name}: the parameter 'largest_segment_id' was not specified, which is necessary for segmentation layers."
-
             segmentation_layer_properties: SegmentationLayerProperties = (
                 SegmentationLayerProperties(
                     **(
@@ -1173,6 +1168,10 @@ class Dataset:
                     chunk_shape = DEFAULT_CHUNK_SHAPE.with_z(1)
                 if chunks_per_shard is None:
                     chunks_per_shard = DEFAULT_CHUNKS_PER_SHARD_ZARR.with_z(1)
+
+            if chunks_per_shard is None and layer.data_format == DataFormat.Zarr3:
+                chunks_per_shard = DEFAULT_CHUNKS_PER_SHARD_FROM_IMAGES
+
             mag_view = layer.add_mag(
                 mag=mag,
                 chunk_shape=chunk_shape,